refactor: Update CTA section on landing page
Improve the design and functionality of the CTA section on the landing page. Adjust the margin-top property to 120px for better spacing. Also, update the border property of the CTA button to have a solid 2px border with the quaternary color. These changes enhance the user experience and make the CTA section more visually appealing.

Signed-off-by: andikatuluspangestu <andhikatuluspangestu@gmail.com>