refactor: Update footer styles and structure
Improve the design and functionality of the footer. Set the background color to #ca4443 and the text color to #FFFFFF. Add padding of 20px and center align the content. Also, update the link color to #FFFFFF and remove the underline. These changes enhance the visual appeal and user experience of the footer.

Signed-off-by: andikatuluspangestu <andhikatuluspangestu@gmail.com>